shiboken6/AutoArrayPointer: Change to use ssize_t
This fits better with Python's Py_ssize_t and Qt's qsizetype,
otherwise, warnings appear:

PySide6/QtWidgets/PySide6/QtWidgets/qgraphicsview_wrapper.cpp:2757:199: warning: ISO C++ says that these are ambiguous, even though the worst conversion for the first is better than the worst conversion for the second:

T &operator[](int pos) { return data[pos]; }
‘operator[](QGraphicsItem**, Py_ssize_t {aka long int})’ <built-in>

+++ sources/shiboken6/libshiboken/helper.h
@@ -84,8 +84,8 @@ class AutoArrayPointer
         AutoArrayPointer &operator=(const AutoArrayPointer &) = delete;
         AutoArrayPointer &operator=(AutoArrayPointer &&) = delete;
 
-        explicit AutoArrayPointer(int size) { data = new T[size]; }
-        T &operator[](int pos) { return data[pos]; }
+        explicit AutoArrayPointer(ssize_t size) { data = new T[size]; }
+        T &operator[](ssize_t pos) { return data[pos]; }
         operator T *() const { return data; }
         ~AutoArrayPointer() { delete[] data; }
     private:
